agentbeats/requirements.txt
a2a-sdk[http-server]==0.3.20
httpx==0.28.1
pydantic==2.12.5
uvicorn==0.38.0
